+++ demo/src/main/java/pe/edu/upc/demo/serviceinterfaces/IUsuarioService.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface IUsuarioService {
     public List<Usuario> list();
     public void insert(Usuario us);
-    public Usuario listId(int id);
+    public Usuario listId(Long id);
     public void update(Usuario u);
-    public void delete(int id);
+    public void delete(Long id);
     public List<String[]> cantidadUsuarioEdadGenero();
     public List<String []> usuariosporgeneroservice();
     public List<String []> mesderegistrosusuarios();
